Clean Up Logic

+++ launcher/src/main.rs
@@ -18,17 +18,17 @@ use crate::archipelago::client::APClient;
 
 #[derive(Clone, Debug)]
 struct LaMulanaConfig {
-    pub version: String,
+    pub version: &'static str,
     pub save_path: String,
-    pub rcd_digest: String,
-    pub dat_digest: String,
-    pub effects_digest: String,
+    pub rcd_digest: &'static str,
+    pub dat_digest: &'static str,
+    pub effects_digest: &'static str,
 }
 
 #[derive(Clone, Debug)]
 struct LaMulanaFileVerification {
     pub file_path: String,
-    pub digest: String
+    pub digest: &'static str
 }
 
 impl LaMulanaFileVerification {
@@ -56,24 +56,31 @@ static VANILLA_SEED_PATH: LazyLock<String> = LazyLock::new(|| { format!("{}vanil
 static BASE_RCD_PATH: &str = "data/mapdata/script.rcd";
 static BASE_DAT_PATH: &str = "data/language/en/script_code.dat";
 static BASE_EFFECTS_PATH: &str = "data/graphics/00/01effect.png";
+
 static VALID_EXE_DIGESTS: LazyLock<HashMap<String, LaMulanaConfig>> = LazyLock::new(|| {
     let home_dir = env::var_os("USERPROFILE").unwrap_or(OsString::new()).into_string().unwrap_or(String::new());
 
     HashMap::from([
-        ("390E26B6A0C1F14BCAC521D2F8E410C4DEAD0B3E2693B2192BD6CA7832CB5B17".to_string(), LaMulanaConfig {
-            version: "1.0.0.1".to_string(),
-            save_path: "save/".to_string(),
-            rcd_digest: "87437780618A3ABDE22BC7200B793FB900169E4F018D1F37D323AC6B5B2F120E".to_string(),
-            dat_digest: "E9F34854D82EBA1E72DD80C573DB1202AA15524FAAC3FC82C8D9F9943BD9F31C".to_string(),
-            effects_digest: "7CB3D2755ECE2E90BC88A81BCEA6C05350E4695182C3798F86F967A5D4BAC466".to_string()
-        }),
-        ("94228016FFFF8A0BA6325140F0CFF6896E2BD0579BB2099D234508DEDE65923F".to_string(), LaMulanaConfig {
-            version: "1.6.6.2".to_string(),
-            save_path: format!("{}/Documents/nigoro/la-mulana/save/", home_dir),
-            rcd_digest: "583DCE2B2BB41E7A1927C6052F7A6AEFEE3F021A792E1AC587E2103C8B5D4CAC".to_string(),
-            dat_digest: "89A2AA21E2CB2DAD6DB5F2EEA474903927980384DE4BC868A9494B1DA3DFED2B".to_string(),
-            effects_digest: "7CB3D2755ECE2E90BC88A81BCEA6C05350E4695182C3798F86F967A5D4BAC466".to_string()
-        })
+        (
+            "390E26B6A0C1F14BCAC521D2F8E410C4DEAD0B3E2693B2192BD6CA7832CB5B17".to_string(),
+            LaMulanaConfig {
+                version: "1.0.0.1",
+                save_path: "save/".to_string(),
+                rcd_digest: "87437780618A3ABDE22BC7200B793FB900169E4F018D1F37D323AC6B5B2F120E",
+                dat_digest: "E9F34854D82EBA1E72DD80C573DB1202AA15524FAAC3FC82C8D9F9943BD9F31C",
+                effects_digest: "7CB3D2755ECE2E90BC88A81BCEA6C05350E4695182C3798F86F967A5D4BAC466"
+            }
+        ),
+        (
+            "94228016FFFF8A0BA6325140F0CFF6896E2BD0579BB2099D234508DEDE65923F".to_string(),
+            LaMulanaConfig {
+                version: "1.6.6.2",
+                save_path: format!("{}/Documents/nigoro/la-mulana/save/", home_dir),
+                rcd_digest: "583DCE2B2BB41E7A1927C6052F7A6AEFEE3F021A792E1AC587E2103C8B5D4CAC",
+                dat_digest: "89A2AA21E2CB2DAD6DB5F2EEA474903927980384DE4BC868A9494B1DA3DFED2B",
+                effects_digest: "7CB3D2755ECE2E90BC88A81BCEA6C05350E4695182C3798F86F967A5D4BAC466"
+            }
+        )
     ])
 });
 
@@ -210,12 +217,8 @@ fn verify_game_files(lm_config: LaMulanaConfig) -> Result<(), String> {
 fn path_exists(file_path: &str, error_if_missing: bool) -> Result<bool, String> {
     match fs::exists(file_path) {
         Ok(exists) => {
-            if error_if_missing {
-                if exists {
-                    Ok(exists)
-                } else {
-                    Err(format!("{} does not appear to exist. Please make sure the launcher is in the base of your La-Mulana install and that the file structure is correct.", file_path))
-                }
+            if error_if_missing && !exists {
+                Err(format!("{} does not appear to exist. Please make sure the launcher is in the base of your La-Mulana install and that the file structure is correct.", file_path))
             } else {
                 Ok(exists)
             }
